place functions in separate js files, remove delete button from post lists, implement switch statement in post-controls
