+++ public/styles/styles.css
@@ -50,6 +50,14 @@
     margin: auto;
 }
 
+.post-preview {
+    background-color: #f0f0f0; /* Light gray */
+    border: #000000 solid 5px; /* Black border */
+    padding: 20px;
+    max-width: 600px;
+    margin: auto;
+}
+
 .post-controls {
     display: flex;
     justify-content: left;
