+++ views/partials/posts.ejs
@@ -49,4 +49,4 @@
         
     </div>
 
-    <%- include('partials/footer') %>
+<%- include('partials/footer') %>
